collbank/requirements.txt
# ===========================================================
# Requirements for COLLBANK
# Date: 7/sep/2023
#
# Python version this works for: 3.8.10 / Ubuntu 20.04.6
#                                3.9 (working)
#
# For a better overview of dependencies: install pipdeptree
# ===========================================================

django==4.2                 # 2.2.28
django-nested-admin==4.1.4  # 3.3.3
django-select2==7.10.1      # 6.3.1
markdown == 3.3.6   
lxml==4.9.3                 # 4.9.1
openpyxl==3.0.9             # 3.0.7
requests==2.26.0
xmltodict                   # Any version ==0.12.0
pytz
pipdeptree==2.23

# =========== OLD definition ==============
# certifi==2021.5.30
# charset-normalizer==2.0.4
# Django==2.2.24
# django-adaptors==0.2.5
# django-nested-admin==3.3.3
# et-xmlfile==1.1.0
# idna==3.2
# lxml==4.6.3
# openpyxl==3.0.7
# python-monkey-business==1.0.0
# pytz==2021.1
# requests==2.26.0
# six==1.16.0
# sqlparse==0.4.1
# urllib3==1.26.6
